@@ -4,18 +4,18 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "naucse_render"
-version = '1.10'
+version = '2.0'
 description = 'Converts course material to naucse.python.cz API'
 readme = "README.md"
-requires-python = ">=3.7"
+requires-python = ">=3.8"
 classifiers = [
     "License :: OSI Approved :: MIT License",
     "Programming Language :: Python :: 3",
 ]
 dependencies = [
     'ansi2html',
-    'mistune',
-    'nbconvert>=6.0.7,<7.0',
+    'mistune>=2.0.3,<4',
+    'nbconvert>=7.0,<8.0',
     'traitlets',
     'click',
     'PyYAML',
